Creacion del crud de cotizacion y arreglo de problemas con la BD

+++ ProyectoDesarrollo/src/Model/Inventory.java
@@ -12,13 +12,13 @@ package Model;
 public class Inventory {
     private String idProducto;
     private String nombreProducto;
-    private int valorUnitario;
+    private float valorUnitario;
     private String descripcion;
     private int lote;
     private int cantidadLote;
     private String idJefe;
 
-    public Inventory(String idProducto, String nombreProducto, int valorUnitario, String descripcion, int lote, int cantidadLote, String idJefe) {
+    public Inventory(String idProducto, String nombreProducto, float valorUnitario, String descripcion, int lote, int cantidadLote, String idJefe) {
         this.idProducto = idProducto;
         this.nombreProducto = nombreProducto;
         this.valorUnitario = valorUnitario;
@@ -46,7 +46,7 @@ public class Inventory {
         this.nombreProducto = nombreProducto;
     }
 
-    public int getValorUnitario() {
+    public float getValorUnitario() {
         return valorUnitario;
     }
